# openapi.yaml
openapi: 3.0.3
info:
  title: Shoe Smart Catalog API (Shore Store)
  description: API for managing the Shore Store catalog, orders, and customer accounts.
  version: 1.0.0
servers:
  - url: http://localhost:2121
    description: Development server
paths:
...
  /api/products:
    get:
      summary: Retrieve products
      description: Get all active products, with optional filtering.
      parameters:
        - name: category
          in: query
          description: Filter shoes by category
          required: false
          schema:
            type: string
            enum: [Sneakers, Formal, Boots, Ladies Wear, Sandals]
        - name: size
          in: query
          description: Filter shoes by available size
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: A list of products.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error.

    post:
      summary: Create a product
      description: Create a new shoe listing. Accepts multipart/form-data.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - price
                - category
                - sizes
                - images
              properties:
                name:
                  type: string
                  example: Retro Runner
                price:
                  type: number
                  example: 35000
                category:
                  type: string
                  enum: [Sneakers, Formal, Boots, Ladies Wear, Sandals]
                  example: Sneakers
                sizes:
                  type: string
                  description: Comma-separated sizes list
                  example: "40,41,42"
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Up to 4 shoe image files
      responses:
        '201':
          description: Product created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Missing fields or images.
        '500':
          description: Server error.
  /api/products/{id}:
    get:
      summary: Retrieve a single product
      description: Get detailed information of a shoe.
      parameters:
        - name: id
          in: path
          required: true
          description: The product ID
          schema:
            type: string
      responses:
        '200':
          description: Product details retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Shoe not found.
        '500':
          description: Server error.
    delete:
      summary: Delete a product (Admin only)
      description: Permanently delete a shoe listing from the database and its images from Cloudinary.
      parameters:
        - name: id
          in: path
          required: true
          description: The product ID
          schema:
            type: string
      responses:
        '200':
          description: Product and associated images deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Shoe listing and associated images deleted successfully
        '403':
          description: Forbidden - Admin access required.
        '404':
          description: Shoe not found.
        '500':
          description: Server error.
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        _id:
          type: string
          example: 60d21b4667d0d8992e610c85
        name:
          type: string
          example: Retro Runner
        price:
          type: number
          example: 35000
        category:
          type: string
          enum: [Sneakers, Formal, Boots, Ladies Wear, Sandals]
          example: Sneakers
        sizes:
          type: array
          items:
            type: integer
          example: [40, 41, 42]
        imageUrls:
          type: array
          items:
            type: string
          example: ["http://res.cloudinary.com/..."]
        isAvailable:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
...